move error code structs to MongoError.swift, define alias for ServerErrorCode, various docs updates
