+++ Sources/MongoSwift/MongoCollection+BulkWrite.swift
@@ -411,68 +411,3 @@ public struct BulkWriteResult {
         }
     }
 }
-
-/// A struct to represent a single write error not resulting from an executed bulk write.
-public struct WriteError: Codable, Equatable {
-    /// An integer value identifying the error.
-    public let code: Int
-
-    /// A description of the error.
-    public let message: String
-
-    private enum CodingKeys: String, CodingKey {
-        case code
-        case message = "errmsg"
-    }
-
-    public static func == (lhs: WriteError, rhs: WriteError) -> Bool {
-        return lhs.code == rhs.code
-    }
-}
-
-/// A struct to represent a write concern error resulting from an executed bulk write.
-public struct WriteConcernError: Codable, Equatable {
-    /// An integer value identifying the write concern error.
-    public let code: Int
-
-    /// A document identifying the write concern setting related to the error.
-    public let details: Document
-
-    /// A description of the error.
-    public let message: String
-
-    private enum CodingKeys: String, CodingKey {
-        case code
-        case details = "errInfo"
-        case message = "errmsg"
-    }
-
-    public static func == (lhs: WriteConcernError, rhs: WriteConcernError) -> Bool {
-        return lhs.code == rhs.code && lhs.details == rhs.details
-    }
-}
-
-/// A struct to represent a write error resulting from an executed bulk write.
-public struct BulkWriteError: Codable, Equatable {
-    /// An integer value identifying the error.
-    public let code: Int
-
-    /// A description of the error.
-    public let message: String
-
-    /// The index of the request that errored.
-    public let index: Int
-
-    /// The request that errored.
-    public let request: WriteModel? = nil
-
-    private enum CodingKeys: String, CodingKey {
-        case code
-        case message = "errmsg"
-        case index
-    }
-
-    public static func == (lhs: BulkWriteError, rhs: BulkWriteError) -> Bool {
-        return lhs.code == rhs.code && lhs.index == rhs.index
-    }
-}
